+++ myceli/src/shipper.rs
@@ -45,6 +45,7 @@ pub struct Shipper<T> {
 }
 
 impl<T: Transport + Send + 'static> Shipper<T> {
+    #[allow(clippy::too_many_arguments)]
     pub fn new(
         storage_path: &str,
         receiver: Receiver<(DataProtocol, String)>,
@@ -53,11 +54,11 @@ impl<T: Transport + Send + 'static> Shipper<T> {
         window_size: u32,
         transport: Arc<T>,
         connected: Arc<Mutex<bool>>,
+        block_size: u32,
     ) -> Result<Shipper<T>> {
         let provider = SqliteStorageProvider::new(storage_path)?;
         provider.setup()?;
-        let storage = Rc::new(Storage::new(Box::new(provider)));
-
+        let storage = Rc::new(Storage::new(Box::new(provider), block_size));
         Ok(Shipper {
             storage,
             window_sessions: BTreeMap::new(),
@@ -95,20 +96,16 @@ impl<T: Transport + Send + 'static> Shipper<T> {
             } => {
                 self.begin_dag_window_session(&cid, &target_addr, retries)?;
             }
-            DataProtocol::RetryDagSession { cid, target_addr } => {
-                if *self.connected.lock().unwrap() && self.window_sessions.contains_key(&cid) {
-                    info!("Received retry dag session, sending get missing req to {target_addr}");
+            DataProtocol::RetryDagSession { cid } => {
+                if *self.connected.lock().unwrap() {
                     if let Some(session) = self.window_sessions.get(&cid) {
-                        let blocks = self.get_dag_window_blocks(&cid, session.window_num)?;
-                        let blocks = blocks.iter().map(|s| s.cid.to_string()).collect();
-                        self.transmit_msg(
-                            Message::DataProtocol(DataProtocol::RequestMissingDagWindowBlocks {
-                                cid: cid.to_string(),
-                                blocks,
-                            }),
-                            &target_addr,
-                        )?;
-                        self.retry_dag_window_session(&cid, &target_addr);
+                        info!(
+                            "Received retry dag session for {cid}, sending get missing req to {}",
+                            &session.target_addr
+                        );
+                        let target_addr = session.target_addr.clone();
+                        self.dag_window_session_run(&cid, session.window_num, &target_addr)?;
+                        self.retry_dag_window_session(&cid);
                     }
                 }
             }
@@ -131,23 +128,28 @@ impl<T: Transport + Send + 'static> Shipper<T> {
             }
             DataProtocol::MissingDagBlocks { cid, blocks } => {
                 if *self.connected.lock().unwrap() {
+                    let target_addr = if let Some(session) = self.window_sessions.get(&cid) {
+                        session.target_addr.to_string()
+                    } else {
+                        sender_addr.to_string()
+                    };
                     // If no blocks are missing, then attempt to move to next window
                     if blocks.is_empty() {
-                        self.increment_dag_window_session(&cid, sender_addr)?;
+                        self.increment_dag_window_session(&cid, &target_addr)?;
                     } else {
                         info!(
                             "Dag {cid} is missing {} blocks, sending again",
                             blocks.len()
                         );
                         for b in blocks.clone() {
-                            self.transmit_block(&b, sender_addr)?;
+                            self.transmit_block(&b, &target_addr)?;
                         }
                         self.transmit_msg(
                             Message::DataProtocol(DataProtocol::RequestMissingDagWindowBlocks {
                                 cid,
                                 blocks,
                             }),
-                            sender_addr,
+                            &target_addr,
                         )?;
                     }
                 }
@@ -194,30 +196,27 @@ impl<T: Transport + Send + 'static> Shipper<T> {
         self.window_sessions.remove(cid);
     }
 
-    fn start_dag_window_retry_timeout(&mut self, cid: &str, target_addr: &str) {
+    fn start_dag_window_retry_timeout(&mut self, cid: &str) {
         let sender_clone = self.sender.clone();
         let cid_str = cid.to_string();
-        let target_addr_str = target_addr.to_string();
+        info!("Starting retry timer at {}", self.retry_timeout_duration);
         let timeout_duration = Duration::from_millis(self.retry_timeout_duration);
         spawn(move || {
             sleep(timeout_duration);
             sender_clone
                 .send((
-                    DataProtocol::RetryDagSession {
-                        cid: cid_str,
-                        target_addr: target_addr_str,
-                    },
+                    DataProtocol::RetryDagSession { cid: cid_str },
                     "127.0.0.1:0".to_string(),
                 ))
                 .unwrap();
         });
     }
 
-    fn retry_dag_window_session(&mut self, cid: &str, target_addr: &str) {
+    fn retry_dag_window_session(&mut self, cid: &str) {
         if let Some(session) = self.window_sessions.get_mut(cid) {
             if session.remaining_window_retries > 0 {
                 session.remaining_window_retries -= 1;
-                self.start_dag_window_retry_timeout(cid, target_addr);
+                self.start_dag_window_retry_timeout(cid);
             }
         }
     }
@@ -247,6 +246,12 @@ impl<T: Transport + Send + 'static> Shipper<T> {
             } else {
                 info!("Dag transfer session for {cid} is complete");
                 self.end_dag_window_session(cid);
+                self.transmit_msg(
+                    Message::ApplicationAPI(messages::ApplicationAPI::DagTransmissionComplete {
+                        cid: cid.to_string(),
+                    }),
+                    target_addr,
+                )?;
             }
         }
         Ok(())
@@ -265,7 +270,7 @@ impl<T: Transport + Send + 'static> Shipper<T> {
             info!("start dag window session for {cid}");
             // Need to reset the window retries here
             self.dag_window_session_run(cid, session.window_num, &session.target_addr)?;
-            self.start_dag_window_retry_timeout(cid, &session.target_addr);
+            self.start_dag_window_retry_timeout(cid);
         }
 
         Ok(())
@@ -290,7 +295,7 @@ impl<T: Transport + Send + 'static> Shipper<T> {
             self.dag_window_session_run(cid, 0, target_addr)?;
             let retries = if retries == 0 { 0 } else { retries - 1 };
             self.open_dag_window_session(cid, retries, target_addr);
-            self.start_dag_window_retry_timeout(cid, target_addr);
+            self.start_dag_window_retry_timeout(cid);
         } else {
             self.open_dag_window_session(cid, retries, target_addr);
         }
@@ -442,6 +447,8 @@ mod tests {
         test_dir: TempDir,
     }
 
+    const BLOCK_SIZE: u32 = 1024 * 3;
+
     impl TestShipper {
         pub fn new() -> Self {
             let mut rng = thread_rng();
@@ -459,7 +466,7 @@ mod tests {
             let db_path = test_dir.child("storage.db");
             let provider = SqliteStorageProvider::new(db_path.path().to_str().unwrap()).unwrap();
             provider.setup().unwrap();
-            let _storage = Rc::new(Storage::new(Box::new(provider)));
+            let _storage = Rc::new(Storage::new(Box::new(provider), BLOCK_SIZE));
             let (shipper_sender, shipper_receiver) = mpsc::channel();
 
             let shipper = Shipper::new(
@@ -470,6 +477,7 @@ mod tests {
                 5,
                 shipper_transport,
                 Arc::new(Mutex::new(true)),
+                BLOCK_SIZE,
             )
             .unwrap();
             TestShipper {