+++ myceli/tests/utils/mod.rs
@@ -13,6 +13,35 @@ use std::sync::Arc;
 use std::thread::{sleep, spawn};
 use transports::{Transport, UdpTransport};
 
+const BLOCK_SIZE: u32 = 1024 * 3;
+
+pub fn wait_receiving_done(receiver: &TestListener, controller: &mut TestController) {
+    let mut prev_num_blocks = 0;
+    let mut num_retries = 0;
+
+    loop {
+        let current_blocks =
+            if let Message::ApplicationAPI(messages::ApplicationAPI::AvailableBlocks { cids }) =
+                controller.send_and_recv(&receiver.listen_addr, Message::request_available_blocks())
+            {
+                cids
+            } else {
+                panic!("Failed to get correct response to blocks request");
+            };
+        let current_num_blocks = current_blocks.len();
+        if current_num_blocks > prev_num_blocks {
+            prev_num_blocks = current_num_blocks;
+            num_retries = 0;
+        } else {
+            if num_retries > 10 {
+                break;
+            }
+            num_retries += 1;
+        }
+        sleep(Duration::from_millis(10));
+    }
+}
+
 pub struct TestListener {
     pub listen_addr: String,
     pub test_dir: TempDir,
@@ -67,9 +96,9 @@ fn start_listener_thread(listen_addr: SocketAddr, db_path: ChildPath) {
         .unwrap();
     transport.set_max_read_attempts(Some(1));
     let transport = Arc::new(transport);
-    let mut listener = Listener::new(&listen_addr, db_path, transport).unwrap();
+    let mut listener = Listener::new(&listen_addr, db_path, transport, BLOCK_SIZE).unwrap();
     listener
-        .start(10, 2)
+        .start(10, 2, BLOCK_SIZE)
         .expect("Error encountered in listener");
 }